fix(vless): align xhttp implementation with maintainer review
