+++ adapter/outbound/vmess.go
@@ -97,18 +97,6 @@ type WSOptions struct {
 	V2rayHttpUpgradeFastOpen bool              `proxy:"v2ray-http-upgrade-fast-open,omitempty"`
 }
 
-type XHTTPOptions struct {
-	Path                 string            `proxy:"path,omitempty"`
-	Host                 string            `proxy:"host,omitempty"`
-	Mode                 string            `proxy:"mode,omitempty"`
-	Headers              map[string]string `proxy:"headers,omitempty"`
-	ScMaxConcurrentPosts int               `proxy:"sc-max-concurrent-posts,omitempty"`
-	ScMaxEachPostBytes   int               `proxy:"sc-max-each-post-bytes,omitempty"`
-	ScMinPostsIntervalMs int               `proxy:"sc-min-posts-interval-ms,omitempty"`
-	NoGRPCHeader         bool              `proxy:"no-grpc-header,omitempty"`
-	XPaddingBytes        string            `proxy:"x-padding-bytes,omitempty"`
-}
-
 func (v *Vmess) StreamConnContext(ctx context.Context, c net.Conn, metadata *C.Metadata) (_ net.Conn, err error) {
 	switch v.option.Network {
 	case "ws":
